requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
urllib3==2.0.5
pytest~=7.1.2
pyyaml~=5.3.1
netmiko~=4.1.0
pexpect
paramiko~=2.11.0
rich~=12.4.4


scp~=0.14.4
tabulate~=0.8.10